backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-phonenumber-field>=8.3.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "phonenumbers>=9.0.16",
    "python-decouple>=3.8",
    "redis>=6.4.0",
    "twilio>=9.8.3",
]
